add Task 41

+++ Task43/Program.cs
@@ -4,13 +4,13 @@
 // b1 = 2, k1 = 5, b2 = 4, k2 = 9 -> (-0,5; -0,5)
 
 Console.Write($"Введите значение b1: ");
-double b1 = Convert.ToInt32(Console.ReadLine());
+double b1 = Convert.ToDouble(Console.ReadLine());
 Console.Write($"Введите значение k1: ");
-double k1 = Convert.ToInt32(Console.ReadLine());
+double k1 = Convert.ToDouble(Console.ReadLine());
 Console.Write($"Введите значение b2: ");
-double b2 = Convert.ToInt32(Console.ReadLine());
+double b2 = Convert.ToDouble(Console.ReadLine());
 Console.Write($"Введите значение k2: ");
-double k2 = Convert.ToInt32(Console.ReadLine());
+double k2 = Convert.ToDouble(Console.ReadLine());
 double x = XCoordinateFinder(b1, k1, b2, k2);
 double y = YCoordinateFinder(b1, k1);
 Console.WriteLine($"Координаты точки пересечения двух прямых: ({x}; {y})");
